add subcmd folder
