+++ Makefile
@@ -3,7 +3,7 @@ all: build test;$(info $(M)...Begin to test and build all of binary.) @
 # Build gitdog binary
 build: ; $(info $(M)...Begin to build gitdog binary.)  @
 	go build -v ./...
-	go build -o bin/gitdog gitdog.go
+	go build -o bin/gitdog src/gitdog.go
 
 # Test
 test: ; $(info $(M)...Begin to test gitdog.)  @
